Add a MetaICP class
This is quite similar in spirit to IncrementalICP,
but matches new clouds to all previously registered clouds
instead of only the last one.
This is computationally more expensive, but provides
a more robust matching when successive clouds do not
alway cover nearby parts of the environment.

As this (just as IncrementalICP) is a standard method to
apply ICP in practice, I expected to find this functionality
in PCL. As it wasn't there yet, I wrote this class.